fix dom failing to update correctly
