+++ packages/react-components/theme-designer/src/ThemeDesigner.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import type { ThemeDesignerProps } from './ThemeDesigner.types';
 import { useStyles } from './ThemeDesigner.styles';
 import { useThemeDesignerReducer } from './useThemeDesignerReducer';
-import { useOverrideReducer } from './useOverrideReducer';
 import { teamsLightTheme, FluentProvider } from '@fluentui/react-components';
 
 import { Nav } from './components/Nav/Nav';
@@ -15,15 +14,14 @@ import { Content } from './components/Content/Content';
 export const ThemeDesigner: React.FC<ThemeDesignerProps> = props => {
   const styles = useStyles();
 
-  const [overrideState, dispatchOverrideState] = useOverrideReducer();
-  const [state, dispatchState] = useThemeDesignerReducer(overrideState);
+  const [state, dispatchState] = useThemeDesignerReducer();
 
   return (
     <FluentProvider theme={teamsLightTheme}>
       <div className={styles.root}>
         <Nav className={styles.nav} brand={state.brand} isDark={state.isDark} overrides={state.overrides} />
         <Sidebar className={styles.sidebar} dispatchState={dispatchState} />
-        <Content className={styles.content} state={state} dispatchOverrideState={dispatchOverrideState} />
+        <Content className={styles.content} state={state} dispatchState={dispatchState} />
       </div>
     </FluentProvider>
   );
