+++ packages/react-components/theme-designer/src/useThemeDesignerReducer.ts
@@ -11,7 +11,6 @@ import { getBrandTokensFromPalette } from './utils/getBrandTokensFromPalette';
 import { brandTeams, brandWeb } from './utils/brandColors';
 
 import type { BrandVariants, Theme } from '@fluentui/react-components';
-import type { OverrideState } from './useOverrideReducer';
 
 export type CustomAttributes = {
   keyColor: string;
@@ -27,6 +26,17 @@ export type DispatchTheme = {
   overrides?: Partial<Theme>;
 };
 
+export type OverrideState = {
+  teamsLight: Partial<Theme>;
+  teamsDark: Partial<Theme>;
+  webLight: Partial<Theme>;
+  webDark: Partial<Theme>;
+  customLight: Partial<Theme>;
+  customDark: Partial<Theme>;
+};
+
+export type DispatchOverride = { type: string; overrides: Partial<Theme> };
+
 export type ReducerState = {
   themeLabel: string;
   brand: BrandVariants;
@@ -35,7 +45,26 @@ export type ReducerState = {
   overrides: Partial<Theme>;
 };
 
-export const useThemeDesignerReducer = (overrideState: OverrideState) => {
+const overrideReducer = (state: OverrideState, action: DispatchOverride) => {
+  switch (action.type) {
+    case 'Teams Light':
+      return { ...state, teamsLight: { ...state.teamsLight, ...action.overrides } };
+    case 'Teams Dark':
+      return { ...state, teamsDark: { ...state.teamsDark, ...action.overrides } };
+    case 'Web Light':
+      return { ...state, webLight: { ...state.webLight, ...action.overrides } };
+    case 'Web Dark':
+      return { ...state, webDark: { ...state.webDark, ...action.overrides } };
+    case 'Custom Light':
+      return { ...state, customLight: { ...state.customLight, ...action.overrides } };
+    case 'Custom Dark':
+      return { ...state, customDark: { ...state.customDark, ...action.overrides } };
+    default:
+      return state;
+  }
+};
+
+export const useThemeDesignerReducer = () => {
   const createCustomTheme = ({
     darkCp,
     hueTorsion,
@@ -54,15 +83,25 @@ export const useThemeDesignerReducer = (overrideState: OverrideState) => {
     };
   };
 
-  const initialState: ReducerState = {
-    themeLabel: 'Teams Light',
-    brand: brandTeams,
-    theme: teamsLightTheme,
-    isDark: false,
-    overrides: {},
-  };
+  const [overrideState, dispatchOverrideState] = React.useReducer(overrideReducer, {
+    teamsLight: {},
+    teamsDark: {},
+    webLight: {},
+    webDark: {},
+    customLight: {},
+    customDark: {},
+  });
 
   const stateReducer = (state: ReducerState, action: DispatchTheme) => {
+    if (action.type === 'Override' && action.overrides) {
+      if (state.themeLabel === 'Custom') {
+        dispatchOverrideState({ type: state.isDark ? 'Custom Dark' : 'Custom Light', overrides: action.overrides });
+      } else {
+        dispatchOverrideState({ type: state.themeLabel, overrides: action.overrides });
+      }
+      action.type = state.themeLabel;
+    }
+
     switch (action.type) {
       case 'Teams Light':
         return {
@@ -114,5 +153,11 @@ export const useThemeDesignerReducer = (overrideState: OverrideState) => {
     }
   };
 
-  return React.useReducer(stateReducer, initialState);
+  return React.useReducer(stateReducer, {
+    themeLabel: 'Teams Light',
+    brand: brandTeams,
+    theme: teamsLightTheme,
+    isDark: false,
+    overrides: {},
+  });
 };
